PWN-4629 - Minor changes
